pyproject.toml
[project]
name = "leanhaul"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "PyYAML>=6.0",
]

[project.scripts]
leanhaul = "leanhaul.cli:main"
